[feat] renamed class repository
